+++ examples/cloud/provider.tf
@@ -8,7 +8,7 @@ terraform {
     }
     elasticstack = {
       source  = "elastic/elasticstack"
-      version = "~>0.3.0"
+      version = "~>0.3.1"
     }
   }
 }