+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Fixed
+- Add new field `allow_custom_routing` in `data_stream` section of [`index_template`](https://www.elastic.co/guide/en/elasticsearch/reference/8.0/indices-put-template.html#put-index-template-api-request-body), which appears only in Elasticsearch version **8.0.0**. Make sure `index_template` resource can work with both **7.x** and **8.x** versions ([#72](https://github.com/elastic/terraform-provider-elasticstack/pull/72))
 
 ## [0.3.0] - 2022-02-17
 ### Added